+++ workalendar/exceptions.py
@@ -7,3 +7,9 @@ class CalendarError(Exception):
     """
     Base Calendar Error
     """
+
+
+class UnsupportedDateType(CalendarError):
+    """
+    Raised when trying to use an unsupported date/datetime type.
+    """