+++ Changelog.md
@@ -3,6 +3,7 @@
 ## master (unreleased)
 
 - Removed dependency to `PyEphem`. This package was the "Python2-compatible" library to deal with the xephem system library. Now it's obsolete, so you don't need this dual-dependency handling, because `ephem` is compatible with Python 2 & Python 3 (#296).
+- Raise an exception when trying to use unsupported date/datetime types (#294).
 
 ## v3.1.1 (2018-11-17)